added address and phone number to Users

+++ app/routers/user.py
@@ -2,7 +2,7 @@ from typing import List
 from fastapi import HTTPException, status, Depends, APIRouter
 from sqlalchemy.orm import Session
 
-from app import models, schemas, utils
+from app import models, oauth2, schemas, utils
 from app.database import get_db
 
 
@@ -45,34 +45,45 @@ async def get_user(email: str, db: Session = Depends(get_db)):
 
 
 @router.delete("/{email}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_user(email: str,  db: Session = Depends(get_db)):
+async def delete_user(email: str,  db: Session = Depends(get_db), current_user=Depends(oauth2.get_current_user)):
     user = db.query(models.User).filter(models.User.email == email).first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"User with email: {email} is not found")
+    if user.id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail=f"Not authorized to delete this user")
     db.delete(user)
     db.commit()
     return user
 
 
 @router.delete("/id/{id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_user_by_id(id: int,  db: Session = Depends(get_db)):
+async def delete_user_by_id(id: int,  db: Session = Depends(get_db), current_user=Depends(oauth2.get_current_user)):
     user = db.query(models.User).filter(models.User.id == id).first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"User with id: {id} is not found")
+    if user.id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail=f"Not authorized to delete this user")
     db.delete(user)
     db.commit()
     return user
 
 
 @router.put("{email}", response_model=schemas.User, status_code=status.HTTP_200_OK)
-async def update_user(email: str, user_data: schemas.UserCreate, db: Session = Depends(get_db)):
+async def update_user(email: str, user_data: schemas.UserCreate, db: Session = Depends(get_db), current_user=Depends(oauth2.get_current_user)):
     user = db.query(models.User).filter(models.User.email == email).first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"User with email: {email} is not found")
+    if user.id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail=f"Not authorized to update this user")
     for field, value in user_data.model_dump().items():
+        if field == "password":
+            value = utils.hash(user_data.password)
         setattr(user, field, value)
     db.commit()
     db.refresh(user)
